+++ esmecata/report/esmecata2taxontology.py
@@ -25,66 +25,6 @@ from esmecata.report.esmecata_compression import RANK2COL
 # CONSTANTS ==========================================================================================================
 
 NCBI = NCBITaxa()
-# RANK2COL = {  # KINGDOM (RED)
-#             'superkingdom': '#592d2d', 'kingdom': '#663535', 'subkingdom': '#7b4040',
-#               # PHYLUM (ORANGE)
-#             'superphylum': '#644125', 'phylum': '#764d2c', 'subphylum': '#835836', 'infraphylum': '#8e603b',
-#               # CLASS (YELLOW)
-#             'superclass': '#635323', 'class': '#736027', 'subclass': '#846f2e', 'infraclass': '#947d36',
-#               # COHORT (KAKHI)
-#             'cohort': '#464b19', 'subcohort': '#555b1f',
-#               # ORDER (GREEN)
-#             'superorder': '#2c4022', 'order': '#344b27', 'suborder': '#425c34', 'infraorder': '#506b41',
-#             'parvorder': '#5a764a',
-#               # FAMILY (TURQUOISE)
-#             'superfamily': '#163036', 'family': '#25454c', 'subfamily': '#32555c',
-#               # TRIBE (TURQUOISE - BLUE)
-#             'tribe': '#2b4759', 'subtribe': '#355468',
-#               # GENUS (BLUE)
-#             'genus': '#243756', 'subgenus': '#354969',
-#               # SECTION (PURPLE)
-#             'section': '#443968', 'subsection': '#534976',
-#               # SERIES (MAGENTA)
-#             'series': '#402c54', 'subseries': '#4f3964',
-#               # SPECIES (PINK)
-#             'species group': '#3e1e36', 'species subgroup': '#4c2643', 'species': '#5b2f50',
-#             'forma specialis': '#713f65', 'subspecies': '#7b476e',
-#               # (PINK - RED)
-#             'varietas': '#502f34', 'subvariety': '#633c41', 'forma': '#71474c', 'serogroup': '#533f41',
-#             'serotype': '#6b4f53', 'strain': '#4c4040', 'isolate': '#685757',
-#               # NO RANK (GREY)
-#             'clade': '#605f5f', 'environmental samples': '#595858', 'incertae sedis': '#616161',
-#             'unclassified': '#706e6e', 'no rank': '#3d3c3c'}
-
-# RANK2COL = {  # KINGDOM (RED)
-#     'superkingdom': '#6f1a12', 'kingdom': '#872015', 'subkingdom': '#a2261a',
-#     # PHYLUM (ORANGE)
-#     'superphylum': '#b26f15', 'phylum': '#c87c14', 'subphylum': '#de8710', 'infraphylum': '#fa960e',
-#     # CLASS (YELLOW)
-#     'superclass': '#c3b21f', 'class': '#d1be14', 'subclass': '#e2cd10', 'infraclass': '#f0db1a',
-#     # COHORT (KAKHI)
-#     'cohort': '#a2af21', 'subcohort': '#c1d02e',
-#     # ORDER (GREEN)
-#     'superorder': '#699924', 'order': '#7ab12a', 'suborder': '#8ecc33', 'infraorder': '#9dde3e', 'parvorder': '#a9e94c',
-#     # FAMILY (TURQUOISE)
-#     'superfamily': '#29b28d', 'family': '#29c198', 'subfamily': '#36d5aa',
-#     # TRIBE (TURQUOISE - BLUE)
-#     'tribe': '#3097a4', 'subtribe': '#3eb5c4',
-#     # GENUS (BLUE)
-#     'genus': '#2a51ba', 'subgenus': '#3863d7',
-#     # SECTION (PURPLE)
-#     'section': '#3a31c6', 'subsection': '#554ce5',
-#     # SERIES (MAGENTA)
-#     'series': '#7838c4', 'subseries': '#904ae3',
-#     # SPECIES (PINK)
-#     'species group': '#883c91', 'species subgroup': '#9c3ea7', 'species': '#af41bb', 'forma specialis': '#c356cf',
-#     'subspecies': '#d66be2',
-#     # (PINK - RED)
-#     'varietas': '#7e2d48', 'subvariety': '#903855', 'forma': '#a44564', 'serogroup': '#bd5979', 'serotype': '#c96685',
-#     'strain': '#d57392', 'isolate': '#e085a2',
-#     # NO RANK (GREY)
-#     'clade': '#585858', 'environmental samples': '#686868', 'incertae sedis': '#828181', 'unclassified': '#939393',
-#     'no rank': '#acacac'}
 
 
 # FUNCTIONS ==========================================================================================================